// frontend/src/app/admin/blogs/models/blog-video-translation.model.ts
// blog-video-translation.model.ts
export class BlogVideoTranslation {
    id?: string;
    video?: string;
    language?: string;
  
    title?: string;
    caption?: string;
    description?: string;
    alt_text?: string;
    aria_label?: string;
    transcript?: string;
  
    seo_title?: string;
    seo_description?: string;
  
    updated_at?: string;
    created_at?: string;

  }